made tyre wear extrapolator a multi file module
